requirements.txt
Pillow==10.1.0
pytest==7.0.1
streamlit==1.28.2
codecarbon==2.3.1
mlflow==2.7.1
protobuf==4.24.4
python-dotenv==1.0.0
scikit_learn==1.3.1
setuptools==68.2.2
tensorflow==2.13.1
uvicorn==0.24.0.post1
fastapi==0.100
python-multipart==0.0.6
